Draft implementation of block begin/end
This is a rough draft for https://github.com/check-spelling/check-spelling/wiki/Feature%3A-Block-Ignore

It includes tests, but not code to trigger it from unknown-words/check-spelling.

Initial file format for `block-delimiters.list`:

     ```block-delimiters.list
     # Description of format 1
     <begin token for format 1>
     <end token for format 1>

     # Description of format 2
     <begin token for format 2>
     <end token for format 2>
     ```

Note that this file format has serious drawbacks and is likely to be
replaced before this feature ships.

Possible replacement file format under consideration:

     ```block-ignore.rules
     name: format 1
     begin-pattern: (regular-expression)
     end-pattern: (regular-expression)
     file-path-pattern: (regular-expression)
     stop-after-pattern: (regular-expression)
     ```