unknown-words: make get_extra_dictionaries output more stable
output written by get_extra_dictionary to stderr is not guaranteed to be
written atomically line at a time to a final stderr.

By writing the stderr output from each task to a distinct file,
it is possible for get_extra_dictionaries to serialize the output and
ensure that it is not interspersed.