upload-sarif: Use single purpose action repository
GitHub doesn't seem to allow repository administrators to specifically
allow using actions that don't live in the root of a repository:
https://support.github.com/ticket/personal/0/3114026

Conversely, it is possible for repository administrators to specifically
allow actions that do live in the root of a repository. As such, even
though this is less than ideal, it's a workable path -- short of
downloading the files as part of the action (which is worth
considering).