unknown-words: Improve handling of test.yml
If a workflow doesn't use `check-spelling@`, look for one that has an id
of `spelling` and treat that as the configuration.

This is mostly to allow `test.yml` to feed `spelling.yml` its config,
but it should make it easier to imagine a standalone config where people
don't want to explicitly use `check-spelling@...`.