apply.t: Fix expired artifact test
Look for an expired artifact in https://github.com/check-spelling-sandbox/autotest-check-spelling,
because it produces new artifacts daily.

Tolerate a repository rename in case the repository moves (it's
possible it'll move to https://github.com/check-spelling/autotest-check-spelling).

Check for an expired artifact on the last page if the first page doesn't
have one.